+++ src/som/compiler/sourcecode_compiler.py
@@ -2,7 +2,8 @@ import os
 from rpython.rlib.streamio import open_file_as_stream
 from rlib.string_stream    import StringStream
 
-from som.compiler.ast.parser import Parser, ParseError
+from som.compiler.ast.parser import Parser
+from som.compiler.parse_error import ParseError
 from som.compiler.class_generation_context import ClassGenerationContext
